Move runner input helpers out of main
